requirements.txt
requests==2.22.0
lxml==4.3.4
beautifulsoup4==4.6.3
PyGithub==1.43.8
tldextract==2.2.0
